checking block timestamps

+++ src/main.cpp
@@ -2784,7 +2784,7 @@ bool ContextualCheckBlockHeader(const CBlockHeader& block, CValidationState& sta
         && block.nVersion.IsLegacy())
         {
           LogPrintf("checking block bits: %08x \n", block.nBits);
-          LogPrintf("block with timestamp: %08x \n", block.nTime);
+          LogPrintf("block with timestamp: %08x \n", (int64_t)block.nTime);
           if(block.nBits != GetNextWorkRequiredLegacy(pindexPrev, &block, consensusParams))
             return state.DoS(100, error("%s: incorrect proof of work at Legacy height %d", __func__, nHeight),
                            REJECT_INVALID, "bad-diffbits");
